Bug fixes in Report, Parasailing module. Reciept format changed as per customer requirements.

+++ src/main/java/com/example/waterSports/controller/RefereeController.java
@@ -2,17 +2,12 @@ package com.example.waterSports.controller;
 
 import com.example.waterSports.modal.ActivityLog;
 import com.example.waterSports.modal.Referee;
-import com.example.waterSports.repo.ActivityLogRepo;
-import com.example.waterSports.repo.ConfigRepo;
-import com.example.waterSports.repo.RefereeRepo;
+import com.example.waterSports.repo.*;
 import com.example.waterSports.utils.Helper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
@@ -20,6 +15,10 @@ import java.util.List;
 @RequestMapping("/ref/")
 public class RefereeController
 {
+    @Autowired
+    OrderWaterSportRepo repoWs;
+    @Autowired
+    OrderParasalingRepo repoPs;
     @Autowired
     RefereeRepo repo;
     @Autowired
@@ -46,6 +45,7 @@ public class RefereeController
     @PostMapping("/")
     public String add(Referee referee)
     {
+        System.out.println("referee = " + referee.toString());
         Referee refSaved = repo.save(referee);
 
         if(referee.getIdOwner() == -1)
@@ -60,15 +60,19 @@ public class RefereeController
     }
 
     @GetMapping("/delete/{id}/")
-    public String delete(@PathVariable Long id)
+    @ResponseBody
+    public Integer delete(@PathVariable Long id)
     {
-        Referee referee = repo.getReferenceById(id);
-        repo.deleteById(id);
+        Integer status = 0;
 
-        repoActivityLog.save(new ActivityLog("Refereee : Recorded Deleted. Referee = " + referee.getName() + ", Owner = " + repo.getReferenceById(referee.getIdOwner()).getName()));
+        if(repo.checkOwner(id) <= 1 && repoWs.countReferences(id) == 0 && repoPs.countReferences(id) == 0)
+        {
+            Referee referee = repo.getReferenceById(id);
+            repoActivityLog.save(new ActivityLog("Refereee : Recorded Deleted. Referee = " + referee.getName() + ", Owner = " + repo.getReferenceById(referee.getIdOwner()).getName()));
+            repo.deleteById(id);
+            status = 1;
+        }
 
-        return "redirect:/ref/";
-
-
+        return status;
     }
 }
